Manifest update: Rust 'debug' branch source

+++ rust/source/debug/Assembly-CSharp/LensDistortion.cs
@@ -48,6 +48,6 @@ public sealed class LensDistortion : PostProcessEffectSettings
 
 	public override bool IsEnabledAndSupported (PostProcessRenderContext context)
 	{
-		return enabled.value && !Mathf.Approximately ((float)intensity, 0f) && ((float)intensityX > 0f || (float)intensityY > 0f) && !RuntimeUtilities.isVREnabled;
+		return enabled.value && !Mathf.Approximately (intensity, 0f) && ((float)intensityX > 0f || (float)intensityY > 0f) && !RuntimeUtilities.isVREnabled;
 	}
 }
